requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
idna==3.4
Nuitka==1.2.3
pycryptodome==3.16.0
pynput==1.7.6
requests==2.28.1
six==1.16.0
urllib3==1.26.13
